pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aegis-ids"
version = "0.3.0b1"
description = "Agentic intrusion detection — detects LLM-driven attacks via cadence analysis and intent chain correlation"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "rich>=13.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "scapy>=2.5",
    "numpy>=1.24",
    "mcp>=1.0",
    "llama-cpp-python>=0.3.0",
]

[project.optional-dependencies]
full = [
    "scikit-learn>=1.3",
    "scipy>=1.11",
]

[project.scripts]
aegis = "aegis.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["aegis*"]
